feat(metastore): dependency checking and writing start / end logs
The first version of dependency checking is implemented with some unit-tests.
The checks are not fully complete yet and they are only for simple 1 on 1
dependencies.

Plus the metastore can now be used to log the start and end of a Tasklog.
The user of the metastore has to keep the startLog objecct around to make
sure that they are paired correctly.