DependencyDAO implementation with some unit tests
DependencyDAO now can save and get Dependencies and will return error info and messages based on
if and why the save operation was not successful.

This relies in part of checking partition-durations which is done in the PartitionDurationTester
which has a few unit tests of its own.

The save with overwrite=false needs more thorough testing and issues with partially set pojo's
also needs thorough testing. There are already quite some unit-tests and tests combining functions
but not all invariants are checked.

The code might be refactored to fit better in a later commit.