First work on Unpartitioned tables + Environment.addTable integration tests
Unpartitioned table support for TableDAO and DependencyDAO was added in which
not having PartitionUnit means a table is unpartitioned, and not having
a transform means it depends on an unpartitioned table.

To make this feature usable, the processDependencies function and return
value + the engine (TableWrapper) + the framemanager, all need to support
it.

This commit also add some integration tests in uhrwerk-engine which test the
loading of configs in combination with persisting tables. But it does not
test running those WrappedTables yet.